refactor(speechlm2): break encoder loading import cycles
